src/main/java/org/pcs/codingsolutions/util/prefixsum/PivotIndexCalculator.java
package org.pcs.codingsolutions.util.prefixsum;

import lombok.AccessLevel;
import lombok.NoArgsConstructor;

import java.util.Arrays;

/**
 * Solution for:
 *
 * Given an array of integers nums, calculate the pivot index of this array.
 *
 * The pivot index is the index where the sum of all the numbers strictly to the left of the index is equal to the sum of all the numbers strictly to the index's right.
 *
 * If the index is on the left edge of the array, then the left sum is 0 because there are no elements to the left. This also applies to the right edge of the array.
 *
 * Return the leftmost pivot index. If no such index exists, return -1.
 */
@NoArgsConstructor(access = AccessLevel.PRIVATE)
public final class PivotIndexCalculator {
    public static int getPivotIndex(int[] nums) {
        int sum = Arrays.stream(nums).sum();
        int leftSum = 0;
        for (int i = 0; i < nums.length; i++) {
            int rightSum = sum - leftSum - nums[i];
            if (leftSum == rightSum) {
                return i;
            }
            leftSum += nums[i];
        }
        return -1; // Time complexity is O(2N) = O(N). Space complexity is O(1)
    }
}
